LayoutTests/platform/ios-simulator/editing/undo/remove-css-property-and-remove-style-expected.txt

This tests removing style attribute after removing CSS property. Undo should bring back the CSS property we removed.

Initially "test" should be bold wrapped with a span:
| <span>
|   style="font-weight: 900;"
|   "test"

Unbolding should remove the CSS style and also remove the span:
| "test"

Undo should bring back both the span and style attribute so that "test" is once again bold:
| "test"

Redo should unbold "test" and there should be no span:
| "test"

Undo once more:
| "test"

Reset, and added id and title:
| <span>
|   id="test_span"
|   style="font-weight: 900;"
|   title="hello, world"
|   "test"

Unbolding should remove the CSS style but shouldn't remove the span:
| <span>
|   id="test_span"
|   style="font-weight: 900;"
|   title="hello, world"
|   "test"

Undo should restore the style attribute and "test" should be bold:
| <span>
|   id="test_span"
|   style="font-weight: 900;"
|   title="hello, world"
|   "test"

Redo should remove the style attribute again:
| <span>
|   id="test_span"
|   style="font-weight: 900;"
|   title="hello, world"
|   "test"

Reset, and added color:blue:
| <span>
|   style="font-weight: 900; color: blue;"
|   "<#selection-anchor>test<#selection-focus>"

Unbolding should remove the font-weight but shouldn't remove the style attribute:
| <span>
|   style="color: blue;"
|   "test"

Undo should reset the style attribute so that "test" is both bold and blue:
| <span>
|   style="color: blue;"
|   "test"

Redo should only remove font-weight and leave "test" blue:
| <span>
|   style="color: blue;"
|   "test"

Setting the forecolor to black should remove both the style attribute and the span:
| <span>
|   style="color: blue;"
|   "test"

Undo should make "test" blue again:
| <span>
|   style="color: blue;"
|   "test"
